ajustes/refatoração

+++ src/components/habits/Habits.jsx
@@ -1,32 +1,10 @@
 import axios from "axios";
 import { useContext, useEffect, useState } from "react";
 import { Buttons, Day, HabitCard, Message, MyHabits, UserHabit, HabitsContainer } from "./HabitsStyle";
-import { API_BASE_URL, weekDays } from './../../mock/data';
+import { API_BASE_URL, weekDays, config } from './../../mock/data';
 import UserContext from "../../context/UserContext";
 import { ThreeDots } from "react-loader-spinner";
-
-function DayCard({ name, daysNumbers, setDaysNumbers, index, loading, days, setDays, isSelected }) {
-    const [selected, setSelected] = useState(isSelected)
-
-    function chooseDay(index) {
-        setSelected(!selected)
-        for(let i = 0; i < daysNumbers.length; i++) {
-            if(daysNumbers[i] === index) {
-                daysNumbers.splice(i, 1)
-                days[index].selected = false
-                setDays([...days])
-                return setDaysNumbers([...daysNumbers])
-            }
-        }
-        days[index].selected = true
-        setDays([...days])
-        setDaysNumbers([...daysNumbers, index])
-    }
-
-    return (
-        <Day disabled={loading} onClick={() => chooseDay(index)} selected={selected}>{name}</Day>
-    )
-}
+import DayCard from "../day-card/DayCard";
 
 export default function Habits() {
     const [habits, setHabits] = useState([])
@@ -42,50 +20,19 @@ export default function Habits() {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [habits])
 
-    function getHabits() {
-        const config = {
-            headers: {
-                "Authorization": `Bearer ${user.token}`
-            }
-        }
-
-        const promise = axios.get(`${API_BASE_URL}/habits`, config)
-        promise.then(res => setHabits(res.data))
-    }
+    const getHabits = () => axios.get(`${API_BASE_URL}/habits`, config(user)).then(res => setHabits(res.data))
+    const deleteHabit = (id) => window.confirm('Deseja deletar este habito ?') ? axios.delete(`${API_BASE_URL}/habits/${id}`, config(user)) : false
 
     function createHabit() {
         setLoading(true)
         const habit = {name, days: daysNumbers.sort((a, b) => a - b)}
-        const config = {
-            headers: {
-                "Authorization": `Bearer ${user.token}`
-            }
-        }
-
-        const promise = axios.post(`${API_BASE_URL}/habits`, habit, config)
+        const promise = axios.post(`${API_BASE_URL}/habits`, habit, config(user))
         promise.then(() => {
-            setLoading(false)
-            setCreating(false)
+            setLoading(false); setCreating(false);
             setDays([...days.map(day => ({ name: day.name, selected: day.selected = false }))])
-            setDaysNumbers([])
-            setName('')
+            setDaysNumbers([]); setName('');
         })
-        promise.catch(res => {
-            alert(`Oops! algo deu errado...${res.response.data}`)
-            setLoading(false)
-        })
-    }
-
-    function deleteHabit(id) {
-        const config = { 
-            headers: {
-                'Authorization': `Bearer ${user.token}`
-            }
-        }
-
-        if(window.confirm('Deseja deletar este habito ?')) {
-            axios.delete(`${API_BASE_URL}/habits/${id}`, config)
-        }
+        promise.catch(res => { alert(`Oops! algo deu errado...${res.response.data}`); setLoading(false);})
     }
 
     return (
@@ -96,10 +43,22 @@ export default function Habits() {
             </MyHabits>
             {creating ? (
                 <HabitCard disable={loading}>
-                    <input type="text" value={name} placeholder="nome do habito" disabled={loading} 
-                        onChange={(e) => setName(e.target.value)} />
+                    <input type="text" value={name} placeholder="nome do habito" disabled={loading} onChange={(e) => setName(e.target.value)} />
                     <div>
-                        {days.map((day, i) => <DayCard key={i} name={day.name} daysNumbers={daysNumbers} days={days} isSelected={day.selected} setDays={setDays} setDaysNumbers={setDaysNumbers} index={i} loading={loading} />)}
+                        {days.map((day, i) => {
+                            return (
+                                <DayCard key={i} 
+                                    name={day.name} 
+                                    daysNumbers={daysNumbers} 
+                                    days={days} 
+                                    isSelected={day.selected} 
+                                    setDays={setDays} 
+                                    setDaysNumbers={setDaysNumbers} 
+                                    index={i} 
+                                    loading={loading} 
+                                />
+                            )
+                        })}
                     </div>
                     <Buttons>
                         <button onClick={() => setCreating(false)} disabled={loading}>Cancelar</button>
@@ -123,9 +82,7 @@ export default function Habits() {
                                             selected = true
                                             counter++
                                         } 
-                                        return (
-                                            <Day key={day.id} selected={selected} readOnly>{day.name}</Day>
-                                        )
+                                        return <Day key={day.id} selected={selected} readOnly>{day.name}</Day>
                                     })}
                                 </div>
                             </div>
